requirements.txt
azure-functions>=1.13.0
requests>=2.28.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
opentelemetry-instrumentation-requests>=0.33b0
dynatrace-opentelemetry-core>=1.313.4